Implement function to find sorted index for number
In this lab you will create a function that returns the lowest index at which a value should be inserted into an array once it has been sorted in ascending order.

Objective: Fulfill the user stories below and get all the tests to pass to complete the lab.

User Stories:

You should have a getIndexToIns function that takes two arguments: an array and a number.
You should use the sort method to sort the array in ascending order.
Your getIndexToIns function should return the lowest index at which the number should be inserted by using the findIndex method.
Your getIndexToIns function should always return a number.
Hint:
The findIndex method is a built-in array method in JavaScript. It takes a callback function and returns the index of the first element that satisfies the condition. Both findIndex and sort are higher-order functions.